test: split demo.py into seperate files and functions
